Delete SDG related projects
Derek stated the reasoning quite well:

[Based on ongoing business development conversations] it seems
like our ability and desire to create SDG demos is currently
outstripping our ability to determine whether and how we can
make money in the SDG space, so we don't know where to focus.
We may know more by the summer, but it'll be difficult to hire
someone without a clearer idea of a potential project.